+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
 - In `Migrate_To_4105._removeBrokenAnnexes`, manage parent's modification date to keep old value because removing an annex
   will `notifyModifiedAndReindex` it's container
 - In `@@item_duplicate_form`, disable annexes if user does not have the permission to `Add annex/Add annexDecision` on future created item
+- Use `OrgaPrettyLinkWithAdditionalInfosColumn` instead `PrettyLinkColumn` in dashboards displaying `persons` and `held_positions`.
+- Added upgrade step to 4106
 
 4.1.23.3 (2020-04-30)
 ---------------------